refactor: split mailbox implementations
